chatbot/backend/requirements.txt
# Core FastAPI framework and dependencies
fastapi>=0.103.1
uvicorn>=0.23.2
sqlalchemy>=2.0.20
pydantic>=2.3.0
python-multipart>=0.0.6

# Authentication and security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1
PyJWT>=2.8.0

# AI Services 
openai>=1.2.0
google-generativeai>=0.3.0
anthropic>=0.5.0

# Database
psycopg2-binary>=2.9.7  # For PostgreSQL
SQLAlchemy-Utils>=0.40.0
pymysql

# Document processing
PyPDF2>=3.0.1
python-docx>=0.8.11
Pillow>=10.0.1
numpy

# HTTP clients
httpx>=0.24.1
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Utilities
starlette>=0.27.0
loguru>=0.7.0
python-magic>=0.4.27
email-validator>=2.0.0

# Testing
pytest>=7.4.2
pytest-asyncio>=0.21.1
